refactor(tooling): extract shared satsuma-core package
