+++ tooling/satsuma-cli/package.json
@@ -22,6 +22,7 @@
     "test": "node --import ./test/setup.js --test test/**/*.test.js"
   },
   "dependencies": {
+    "@satsuma/core": "file:../satsuma-core",
     "commander": "^14.0.3",
     "web-tree-sitter": "^0.26.7"
   },
