+++ tooling/satsuma-core/package.json
@@ -1,0 +1,21 @@
+{
+  "name": "@satsuma/core",
+  "version": "0.1.0",
+  "private": true,
+  "description": "Shared Satsuma core utilities",
+  "license": "MIT",
+  "type": "module",
+  "exports": {
+    ".": "./dist/index.js",
+    "./format": "./dist/format.js",
+    "./types": "./dist/types.js"
+  },
+  "types": "./dist/index.d.ts",
+  "scripts": {
+    "build": "tsc",
+    "prepare": "tsc"
+  },
+  "devDependencies": {
+    "typescript": "^6.0.2"
+  }
+}
