+++ tooling/satsuma-core/src/index.ts
@@ -1,0 +1,2 @@
+export { format } from "./format.js";
+export type { SyntaxNode, Tree } from "./types.js";
