+++ tooling/satsuma-core/src/types.ts
@@ -1,0 +1,20 @@
+export interface SyntaxNode {
+  type: string;
+  text: string;
+  isNamed: boolean;
+  children: SyntaxNode[];
+  namedChildren: SyntaxNode[];
+  childCount: number;
+  child(index: number): SyntaxNode | null;
+  childForFieldName?(name: string): SyntaxNode | null;
+  parent: SyntaxNode | null;
+  startPosition: { row: number; column: number };
+  endPosition: { row: number; column: number };
+  startIndex: number;
+  endIndex: number;
+  isMissing: boolean;
+}
+
+export interface Tree {
+  rootNode: SyntaxNode;
+}
