+++ tooling/vscode-satsuma/server/package.json
@@ -12,6 +12,7 @@
     "pretest": "tsc"
   },
   "dependencies": {
+    "@satsuma/core": "file:../../satsuma-core",
     "vscode-languageserver": "^9.0.1",
     "vscode-languageserver-textdocument": "^1.0.12",
     "web-tree-sitter": "^0.26.7"
